utils/check-image-resolution/pyproject.toml
[project]
name = "check-image-resolution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
]
